Add user vuex modules
