+++ kolibri/plugins/user/assets/src/constants.js
@@ -4,3 +4,9 @@ export const PageNames = {
   SIGN_UP: 'SIGN_UP',
   PROFILE: 'PROFILE',
 };
+
+export const pageNameToModuleMap = {
+  [PageNames.SIGN_IN]: 'signIn',
+  [PageNames.SIGN_UP]: 'signUp',
+  [PageNames.PROFILE]: 'profile',
+};
