+++ kolibri/plugins/user/assets/src/modules/profile/handlers.js
@@ -1,0 +1,8 @@
+import { PageNames } from '../../constants';
+
+export function showProfilePage(store) {
+  store.dispatch('resetAndSetPageName', {
+    pageName: PageNames.PROFILE,
+  });
+  store.commit('profile/RESET_STATE');
+}
