+++ kolibri/plugins/user/assets/src/modules/signIn/index.js
@@ -1,0 +1,14 @@
+export default {
+  namespaced: true,
+  state: {
+    hasMultipleFacilities: null,
+  },
+  mutations: {
+    SET_STATE(state, payload) {
+      Object.assign(state, payload);
+    },
+    RESET_STATE(state) {
+      state.hasMultipleFacilities = null;
+    },
+  },
+};
