--- kolibri/plugins/user/assets/src/state/pluginModule.js
@@ -1,12 +1,0 @@
-import mutations from './mutations';
-import * as actions from './actions';
-
-export default {
-  state: {
-    pageName: undefined,
-    pageState: {},
-    facilityId: undefined,
-  },
-  actions,
-  mutations,
-};
