+++ kolibri/plugins/user/assets/test/makeStore.js
@@ -1,5 +1,5 @@
 import { coreStoreFactory } from 'kolibri.coreVue.vuex.store';
-import pluginModule from '../src/state/pluginModule';
+import pluginModule from '../src/modules/pluginModule';
 
 export default function makeStore() {
   return coreStoreFactory(pluginModule);
