--- kolibri/plugins/user/assets/test/util/makeStore.js
@@ -1,6 +1,0 @@
-import { coreStoreFactory } from 'kolibri.coreVue.vuex.store';
-import pluginModule from '../../src/state/pluginModule';
-
-export default function makeStore() {
-  return coreStoreFactory(pluginModule);
-}
