+++ kolibri/plugins/user/assets/test/views/user-page.spec.js
@@ -4,8 +4,6 @@ import makeStore from '../makeStore';
 
 jest.mock('kolibri.urls');
 
-jest.mock('kolibri.urls');
-
 function makeWrapper() {
   return mount(UserPage, {
     store: makeStore(),
